requirements.txt
numpy==1.23.5
streamlit==1.22.0
Pillow==9.5.0
keras==2.12.0
tensorflow==2.12.0
Django
python-dotenv==0.21.0
django-ckeditor==6.5.1
django-ckeditor-5==0.2.2
django-crispy-forms==1.14.0
django_debug_toolbar==3.8.1
django-widget-tweaks==1.4.12
django-dotenv==1.4.2
django-environ==0.9.0
django-social-share==2.3.0
django-static-fontawesome==5.15.4.0
django-taggit==3.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
whitenoise==6.2.0
